Improve the score inspector interface
Separate the inspector into more tabs (notation, audio, MIDI
and advanced), so that it takes up less space and so that the
options are more logically organised. Only have the bare
minimum on the default tab (the notation tab) to avoid confusing
anyone who doesn't want to edit the advanced options.

Also improve the overall experience by making the audio options
work the same way as the MIDI options.

Bug: T114227
Change-Id: I9b927ffd0e04674ac8a6f27aaf7219eab307f813