Enable shell restrictions for all binaries
If firejail is available and configured, MediaWiki will now run
lilypond, abc2ly, timidity, and imagemagick inside a firejail with the
default restrictions and no network access. In addition, this moves away
from the deprecated wfShellExec() in favor of the newer
Shell::command().

This affects output on Special:Version and in articles.

I used <https://en.wikipedia.org/wiki/Pastime_with_Good_Company> as the
score to test with.

Bug: T181535
Depends-On: Icb29d4c48ae7f92fb5635e3865346c98f47abb01
Change-Id: I300fc6ed96ca8c8cf2c16644871c9eb7724b915d