Check for PNG metadata before generating HTML
When embedding the PNG file(s), the dimensions of the file are
placed into the `<img` tag, being loaded from a metadata file. Only
the existance of the file is checked, and not if it has valid values.
This change adds a check for if the values exist, and if not, to
regenerate the images and metadata.

Bug: T143070
Change-Id: I8dff67269a6cdda555aab33bf0f6c5f4b44c12a4