Remove rawAndMidi from cache options
This was an unnecessary addition from I6a0a239b6. It is un-needed as
whenever the `midi` block is included in a `<score` tag, it will be
generated. In raw scores, if the attributes have been changed to include
the files, either the MIDI would already exist, thus no need for
re-generation, or the code would be modified, making the hash different.

Bug: T49448
Change-Id: I56c6034834a5b706c2f897362212e0ad62314b0c