Regression fix for TMH + Score integration
If there wasn't another TMH invocation on the page, the audio player
generated by a <score vorbis=1>...</score> would not invoke the TMH
frontend, resulting in inconsistent appearance and failure to run
the Ogg output on Safari/IE/Edge.

This was due to changes in the frontend code for TMH a while back;
added a workaround to invoke the TMH-side code that sets it up, but
not sure this is nice. :D

Still want a better interface between these two for the future.

Bug: T148716
Change-Id: I0b86d37ac268a45dd3417fc7cbb49f3f7d3adbee