Use different MIME type for OGG files
The MIME type that was used, `audio/ogg` is not supported by TMH, and so
`application/ogg` has to be used instead. This gets the length
calculation and the repeat functions working again. In addition, the
length of the audio files was pointing to the full hash of the code, and
not the proper file name. This commit fixes these bugs.

Bug: T192550
Bug: T200835
Change-Id: Ia14d208852af0d56bb4b4d2f4b251d4cdf5c1edd