Check for override_midi in TMH player
The current incorrect behaviour is that the audio player (TMH) takes the
length of the audio with the name of the score, which is incorrect when
override_midi is used, as in that case, the hash of the file is used.

This also adds the length check for non-overridden audio files.

Bug: T203560
Change-Id: Ie10850529df50e2e923df00932a4633b9c130111