Remove incorrect paper variables
raggedright and raggedbottom were renamed to ragged-right and
ragged-bottom in LilyPond 2.7.32 [1], and the old names are ignored
since then (there is no backwards compatibility). This means that these
variables have never had an effect (the initial commit of the Score
extension was over five years after the release of LilyPond 2.7.32);
whatever their intended effect was at the time, adding them now (by
fixing the names) would be a breaking change to many existing scores, so
let’s just remove them to avoid confusion.

[1]: http://lilypond.1069038.n5.nabble.com/LilyPond-2-7-32-is-out-td130222.html

Change-Id: I953cd091eafd7e34e84b1e2baea1f26c881a60c1