Add tracking categories to pages overriding files
As a step towards deprecation of the override_midi and override_audio
attributes of scores, this commit adds a tracking category for those
pages.

The parameters override_midi and override_audio are also no longer
needed in favour of manually adding a thumbnail audio file outside of
the score tag.

Bug: T135597
Change-Id: Ib448042c6dce2fd07a431ceac15d16b05aa958da