Turn off .ly source downloads
Hide the UI for offering .ly files for download behind a config option,
off by default. If the user renders such a file using a local LilyPond
install, malicious scripts may be executed. It's like offering a .exe
download. Maybe we could allow it if we could find a way to scan the
file for malware.

Change-Id: If2114a368a1669fd6acc7a69a835c18bcbbdc9cd