Drop TimedMediaHandler support
mwembed has been disabled by default, and can't be explicitly
selected from the caller. VideoJS has poor audio support, and
<audio controls> is apparently superior. So use <audio> directly. To
avoid the need for client-side Ogg support, use MP3 unconditionally,
which has been patent-free since 2017 and is supported on all browsers.

The <audio> tag apparently cannot use a duration hint, and TMH was
previously being used to calculate the audio duration. Keep storing it
anyway to make migration to a new player in future easier. Estimate the
duration from the WAV file size instead of using TMH. Specify 44100 as
the FluidSynth sample rate so that this method will continue to be
correct even if the FluidSynth default sample rate changes.

TMH put a width attribute on the <audio> tag, but this is invalid. We
don't need to specify the audio player width since there's no way to
control it.

Bug: T226910
Bug: T100106
Change-Id: Id85929c8d2a20457b252a0505e854bc7b05b03bc