Allow setting a different path for `convert` just for Score
While transitioning Score to Shellbox, we want the default `convert`
command used by core and other places to still use our firejail wrapper.
However that wrapper doesn't exist in the Shellbox container, so we just
need to use a plain `/usr/bin/convert` for Score purposes.

Introduce a new config option, $wgScoreImageMagickConvert, which can be
used to override the default $wgImageMagickConvertCommand. By default
it is set to false, so it'll fallback to the core setting.

Change-Id: I9004e89928cf35b3ef07e7a475ba897214f46497