Increase lilypond version cache TTL to 1 hour
The lilypond version is used for the rendered image cache key and
displayed on Special:Version. It's cached on the local server (APCu) for
60s.

Most traffic to Shellbox is just executing `lilypond --version`. Given
that we rarely change the lilypond version, we should be able to
increase the cache TTL to an hour with no major negative effects.

Whenever lilypond is upgraded in the Shellbox container, it'll take up
to an hour for all servers to realize they need to re-render previously
cached images. I'm not sure this is a bad thing since it would spread
out the invalidation instead of a stampede within a minute of deploying.

I think it's probably OK if Special:Version is out of date for an hour
as well.

Change-Id: Id6f5d78be591b49a2b39cc4641d2250c1fabec24