Record shell outs in statsd
For better visibility into how often Score is needing to talk to
Shellbox, and for what type of request. In theory we could try
to track these on the Shellbox side but we don't yet have statsd
set up there and it would need to do some command normalization...
so easier to do in Score.

Change-Id: Ic219aa5d81a27af2b0e2c601e98f445d1dbf2014