Fix score-audio-alt message showing up if <score> preceded by colon
The "Audio playback is not supported in your browser" message is
supposed to be hidden by the browser when the file in <audio> is
playable. But the legacy PHP Parser which operates on lines rather
than HTML structure got confused by the extra newline and moved the
<div> containing the error message outside of the <audio> tag so
it would appear unconditionally.

Removing the newline fixes this issue. Credit to Arlolra for
figuring this out.

Bug: T266307
Change-Id: I5f75955945aef5ebd58d8cf12271a829c56940a7