Use a ScopedCallback to simplify cleaning up the $factoryDirectory
Using a ScopedCallback will ensure the cleanup is done even if an
exception is thrown, so we can remove the extra level of indentation
needed to try/catch so cleanup happens.

While we're at it, rename eraseFactory to eraseDirectory so it matches
createDirectory.

This is based on CR discussion of I5ab357f17617 in EasyTimeline.

Change-Id: Ia818790731bf1727333a971dc17cb72085b8ff2d