Don't configure parser-factory-callback
Configuring parser-factory-callback is not necessary here because it's
the same as the default for the string value type. Setting a data type
specific custom parser will soon allow for data type specific overrides
of the parser whenever a statement value is deserialized, and setting
it may result in additional data type lookups.

Bug: T360603
Change-Id: I60c3f3422934c87e9e01110f6b924be7de1a2234