Close dialog without confirmation if the user hasn't interacted
Recently we changed ve.ui.MWExtensionDialog to add `hasMeaningfulEdits`
which determines whether the user has interacted with a dialog and thus
allows the dialog to be closed without confirmation if they have made
no changes.

This change makes sure that the opening state is stored so that we can
detect this case.

Bug: T334513
Change-Id: I3f65551a4a58f0c4df2dbd785e9fce63b56766bd